Move test prepare logic to base configuration
