+++ wdio.conf.base.js
@@ -1,3 +1,4 @@
+import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { getTestTempDirectory } from './test/support/utils.js';
@@ -8,6 +9,7 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 global.tempDirectory = path.join(__dirname, 'temp');
 
+// Configuration for all testable browsers.
 let capabilities = [
   {
     browserName: 'chrome',
@@ -105,4 +107,18 @@ export const config = {
     ui: 'bdd',
     timeout: 60_000
   },
+
+  async onPrepare (_config, capabilities) {
+    // Ensure we have a clean temp directory.
+    await fs.rm(global.tempDirectory, { recursive: true, force: true });
+    await fs.mkdir(global.tempDirectory);
+
+    // And a subdirectory for each test environment -- they run in parallel,
+    // so need their own isolated space to play in.
+    for (const capability of capabilities) {
+      const browserDirectory = getTestTempDirectory(capability);
+      await fs.rm(browserDirectory, { recursive: true, force: true });
+      await fs.mkdir(browserDirectory);
+    }
+  }
 };
